blockchain.py
import os
import json
from koii import Finnie, TaskNode, ArweaveStorage, AttentionOracle

class QuantumBlockchain:
    def __init__(self):
        self.finnie = Finnie()
        self.arweave = ArweaveStorage()
        self.attention = AttentionOracle()
        self.config = TerminusaConfig()

    async def initialize_player(self, username):
        if not self._validate_username(username):
            raise ValueError("Invalid username format")
            
        wallet = await self.finnie.connect()
        return {
            'username': username,
            'address': wallet['address'],
            'tac': self.config.INITIAL_TAC,
            'koii': 0,
            'inventory': {}
        }

class ArweaveStorage:
    def __init__(self):
        self.gateway = os.getenv('ARWEAVE_GATEWAY', 'https://arweave.net')
        
    async def store(self, data):
        tx = Transaction(data=json.dumps(data))
        await tx.sign()
        await tx.post()
        return tx.id